+++ app/Http/Controllers/SuperAdmin/AccountController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers\SuperAdmin;
+
+use App\Http\Controllers\Controller;
+use Illuminate\View\View;
+
+class AccountController extends Controller
+{
+    public function edit(): View
+    {
+        return view('superadmin.account.edit');
+    }
+}